+++ UnsplashClient/app/build.gradle.kts
@@ -67,4 +67,6 @@ dependencies {
 
     implementation(libs.hilt.android)
     ksp(libs.hilt.android.compiler)
+
+    implementation(libs.androidx.navigation.compose)
 }
