+++ UnsplashClient/app/src/main/java/com/example/unsplashclient/presentation/MainActivity.kt
@@ -1,0 +1,41 @@
+package com.example.unsplashclient.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.ui.Modifier
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.example.unsplashclient.presentation.ui.theme.UnsplashClientTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            UnsplashClientTheme {
+                Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+                    val navController = rememberNavController()
+                    NavHost(
+                        navController = navController,
+                        startDestination = ScreenRoute.SearchScreen.route,
+                        modifier = Modifier.padding(innerPadding)
+                    ) {
+                        composable(route = ScreenRoute.SearchScreen.route) {
+                            // TODO
+                        }
+                        composable(route = ScreenRoute.PhotoDetailScreen.route) {
+                            // TODO
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
