+++ UnsplashClient/app/src/main/java/com/example/unsplashclient/presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.unsplashclient.ui.theme
+package com.example.unsplashclient.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
