+++ UnsplashClient/app/src/main/java/com/example/unsplashclient/presentation/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.example.unsplashclient.ui.theme
+package com.example.unsplashclient.presentation.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
