+++ UnsplashClient/gradle/libs.versions.toml
@@ -3,7 +3,7 @@ agp = "8.9.0"
 converterMoshi = "2.11.0"
 hiltAndroid = "2.56.1"
 kotlin = "2.0.21"
-ksp = "1.9.0-1.0.13"
+ksp = "2.0.21-1.0.25"
 coreKtx = "1.15.0"
 junit = "4.13.2"
 junitVersion = "1.2.1"
@@ -12,10 +12,12 @@ lifecycleRuntimeKtx = "2.8.7"
 activityCompose = "1.10.1"
 composeBom = "2024.09.00"
 moshiKotlin = "1.15.1"
+navigationCompose = "2.9.0"
 retrofit = "2.11.0"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
+androidx-navigation-compose = { module = "androidx.navigation:navigation-compose", version.ref = "navigationCompose" }
 converter-moshi = { module = "com.squareup.retrofit2:converter-moshi", version.ref = "converterMoshi" }
 hilt-android = { module = "com.google.dagger:hilt-android", version.ref = "hiltAndroid" }
 hilt-android-compiler = { module = "com.google.dagger:hilt-android-compiler", version.ref = "hiltAndroid" }
